third_party/WebKit/LayoutTests/http/tests/appcache/obsolete-error-events-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Test that master list entries receive errors when manifest becomes obsolete.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eventDetailsFromWindow.type is "obsolete"
PASS eventDetailsFromFrame.type is "error"
PASS eventDetailsFromFrame.reason is "manifest"
PASS eventDetailsFromFrame.url is ""
PASS eventDetailsFromFrame.status is 404
PASS successfullyParsed is true

TEST COMPLETE
